requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
joblib>=1.1.0
streamlit>=1.0.0
python-pptx>=0.6.21